[#148] :sparkles: refactor: wonFormat 유틸함수 helper 에서 utils 로 경로 이동
